Backend/requirements.txt
asgiref==3.8.1
autopep8==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyaml==24.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
